@@ -9,42 +9,35 @@
     "build": "nuxt build",
     "start": "cross-env NODE_ICU_DATA=node_modules/full-icu nuxt start",
     "generate": "nuxt generate",
-    "lint": "eslint --ext .js,.vue --ignore-path .gitignore .",
-    "precommit": "npm run lint"
+    "lint": "eslint --ext .js,.vue --ignore-path .gitignore ."
   },
   "dependencies": {
     "@nuxtjs/axios": "^5.5.4",
-    "@nuxtjs/pwa": "^2.6.0",
+    "@nuxtjs/pwa": "^3.0.0-0",
     "axios": "^0.19.0",
     "bootstrap": "^4.1.3",
     "braintree-web": "^3.47.0",
     "cross-env": "^5.2.0",
     "full-icu": "^1.3.0",
     "lodash": "^4.17.14",
-    "nuxt": "^2.8.1",
     "nuxt-i18n": "^5.12.7",
+    "nuxt": "^2.10.2",
     "paypal-checkout": "^4.0.280",
     "standard-slugify": "^1.0.0"
   },
   "devDependencies": {
-    "@nuxtjs/eslint-config": "^0.0.1",
-    "babel-eslint": "^8.2.1",
-    "eslint": "^5.0.1",
-    "eslint-config-prettier": "^3.1.0",
-    "eslint-config-standard": "^13.0.0",
-    "eslint-import-resolver-alias": "^1.1.2",
-    "eslint-loader": "^2.2.1",
-    "eslint-plugin-import": "^2.18.0",
-    "eslint-plugin-jest": "^22.7.2",
-    "eslint-plugin-node": ">=7.0.1",
-    "eslint-plugin-prettier": "2.6.2",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-standard": ">=4.0.0",
-    "eslint-plugin-vue": "^5.2.3",
-    "node-sass": "^4.12.0",
-    "nodemon": "^1.19.1",
-    "prettier": "1.14.3",
-    "sass-loader": "^7.1.0",
+    "@nuxtjs/eslint-config": "^1.0.1",
+    "@nuxtjs/eslint-module": "^1.0.0",
+    "@nuxtjs/google-analytics": "^2.2.1",
+    "@nuxtjs/style-resources": "^1.0.0",
+    "babel-eslint": "^10.0.1",
+    "eslint": "^6.1.0",
+    "eslint-config-prettier": "^4.1.0",
+    "eslint-plugin-nuxt": ">=0.4.2",
+    "eslint-plugin-prettier": "^3.0.1",
+    "node-sass": "^4.13.0",
+    "prettier": "^1.16.4",
+    "sass-loader": "^8.0.0",
     "vue-waypoint": "^3.2.2"
   }
 }
